trading-room/requirements.txt
yfinance>=0.2.40
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0
requests>=2.31.0
pytest>=7.4.0
rich>=13.7.0
